# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Webtableau API
  description:
    An API that allows various operations related to receiving and sending data of recipe instructions.
  version: 0.1-TEST
servers:
  - url: http://localhost:8080
tags:
  - name: getrecipedata
    description: Operations related to receive recipe data.
  - name: postrecipedata
    description: Operations related to post recipe data.
  - name: putrecipedata
    description: Operations related to update recipe data.
  - name: getuserdata
    description: Operations related to receive user data.
  - name: postuserdata
    description: Operations related to posting user data.
paths:
  /recipe:
    get:
      summary: Get a recipe.
      tags:
        - getrecipedata
      operationId: getrecipe
      parameters:
        - in: query
          name: id
          description: Loading recipe.
          required: true
          schema:
            type: integer
        - in: query
          name: uid
          description: User-ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search for specific recipe.
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Recipe'
    post:
      summary: Add a recipe to the database.
      tags:
        - postrecipedata
      operationId: postrecipe
      parameters:
        - in: query
          name: uid
          description: User-ID
          required: true
          schema:
            type: string
      requestBody:
        description: Add a new recipe.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: Added recipe to the database.
        '400':
          description: Something went wrong.
    put:
      summary: Update a recipe.
      tags:
        - putrecipedata
      operationId: putrecipe
      parameters:
        - in: query
          name: id
          description: Loading recipe.
          required: true
          schema:
            type: integer
        - in: query
          name: uid
          description: User-ID
          required: true
          schema:
            type: string
      requestBody:
        description: Add a new recipe.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: Added recipe to the database.
        '400':
          description: Something went wrong.
    delete:
      description: Deletes a recipe by recipe id and user uid.
      operationId: deleterecipe
      parameters:
        - in: query
          name: recipeid
          description: Recipe id for deleting a selected recipe.
          required: true
          schema:
            type: integer
        - in: query
          name: uid
          description: Uid of user, who wants to delete a selected recipe.
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Deleted recipe from database.
        '400':
          description: Something went wrong.



  /recipe/page:
    get:
      summary: Get a list of all recipes.
      tags:
        - getrecipedata
      operationId: getrecipepage
      x-spring-paginated: true
      parameters:
        - in: query
          name: page
          description: The page number to retrieve.
          required: true
          schema:
            type: integer
        - in: query
          name: size
          description: The number of items per page.
          required: true
          schema:
            type: integer
        - in: query
          name: title
          description: A title or text someone is looking for.
          required: false
          schema:
            type: string
        - in: query
          name: uid
          description: The uid of a user.
          required: true
          schema:
            type: string
        - in: query
          name: created
          description: only filter for user recepie.
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of all available recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRecipes'
  /notification:
    post:
      summary: Sends a notification.
      operationId: postnotification
      parameters:
        - in: query
          name: uid
          description: Uid of user.
          required: true
          schema:
            type: string
        - in: query
          name: token
          description: TOKEN.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /favourites:
    get:
      summary: Get a list of all favourite recipes.
      tags:
        - getrecipedata
      operationId: getfavouriterecipes
      x-spring-paginated: true
      parameters:
        - in: query
          name: uid
          description: The uid of a user, to find favourite recipes.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of all favourite recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFavourites'
    post:
      summary: Add a favourite to a User and get a list of all favourite recipes.
      tags:
        - postrecipedata
      operationId: postfavouritrecipes
      x-spring-paginated: true
      parameters:
        - in: query
          name: uid
          description: The uid of a user, to find favourite recipes.
          required: true
          schema:
            type: string
        - in: query
          name: recipeid
          description: The uid of a recipe.
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: List of all favourite recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFavourites'
    delete:
      summary: Delete a favourite of a User and get a list of all favourite recipes.
      tags:
        - getrecipedata
      operationId: deleltefavouriterecipes
      x-spring-paginated: true
      parameters:
        - in: query
          name: uid
          description: The uid of a user, to find favourite recipes.
          required: true
          schema:
            type: string
        - in: query
          name: recipeid
          description: The recipe page id of a recipe, what someone wants to delete.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of all favourite recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFavourites'

components:
  schemas:
    Recipe:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdby:
          type: string
        shortsummary:
          type: string
        preparation:
          type: string
        ingredients:
          type: string
        bakingtime:
          type: integer
        picture:
          type: string
    RecipeShort:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        shortsummary:
          type: string
        picture:
          type: string
        bakingTime:
          type: integer
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        username:
          type: string
        profilepicture:
          type: string
        favourites:
          type: string
    PaginatedRecipes:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RecipeShort'
        currentpageNo:
          type: integer
        pagesize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        islast:
          type: boolean
    PaginatedFavourites:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RecipeShort'